Whatsapp-Chat-Analyzer-main/Whatsapp-Chat-Analyzer-main/requirements.txt
# Core Framework
streamlit==1.31.0
python-dotenv==1.0.0

# Database & ORM
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
bcrypt==4.1.2

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Graph Analytics
networkx==3.2.1
matplotlib==3.8.2
streamlit-agraph==0.0.45
plotly==5.18.0

# AI & Embeddings
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.16
langchain-google-genai==0.0.6
google-generativeai==0.3.2
sentence-transformers==2.3.1
faiss-cpu

# Text Analysis
wordcloud==1.9.3
emoji==2.10.0
textblob==0.17.1
nltk==3.8.1

# Utilities
python-dateutil==2.8.2
regex==2023.12.25
